Fixed gcc compilation error (x64).

+++ compatible_edition/include/st_asio_wrapper_tcp_socket.h
@@ -91,7 +91,7 @@ protected:
 			return;
 		}
 
-		int loop_num = GRACEFUL_CLOSE_MAX_DURATION * 100; //seconds to 10 milliseconds
+		ssize_t loop_num = GRACEFUL_CLOSE_MAX_DURATION * 100; //seconds to 10 milliseconds
 		if (sync)
 		{
 			while (--loop_num >= 0 && is_closing())
@@ -143,7 +143,7 @@ protected:
 		case 9:
 			if (is_closing())
 			{
-				int loop_num = reinterpret_cast<int>(user_data);
+				ssize_t loop_num = reinterpret_cast<ssize_t>(user_data);
 				--loop_num;
 
 				if (loop_num > 0)
